+++ packages/orq-rc/RELEASES.md
@@ -10629,3 +10629,13 @@ Based on:
 - [python v4.15.0-rc.15] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.15] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.15 - packages/orq-rc
+
+## 2026-09-03 07:48:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.3 (2.934.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.16] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.16 - packages/orq-rc